Introduce AVOID_UB macro + configure.py --all-code and -D flags

+++ src/dlls/objects/496_snowhorn/snowhorn.c
@@ -977,7 +977,7 @@ void dll_496_func_1D68(Object* self, SnowHorn_Data* objdata, SnowHorn_Setup* set
             setup = (SnowHorn_Setup*)self->setup;
             if (frostWeed && frostWeed->id == 0x3FB && vec3_distance_xz_squared(&self->positionMirror, &frostWeed->positionMirror) < setup->unkRadius * setup->unkRadius) {
                 if (!((DLL_227_Tumbleweed*)frostWeed->dll)->vtbl->func11(frostWeed)) {
-                    ((DLL_227_Tumbleweed*)(frostWeed->dll))->vtbl->func10(frostWeed, (s32)&objdata->playerPositionCopy);
+                    ((DLL_227_Tumbleweed*)(frostWeed->dll))->vtbl->func10(frostWeed, &objdata->playerPositionCopy);
                     objdata->frostWeed = frostWeed;
                     if (0){
                         objdata->garundaTe_weedsEaten = GARUNDA_TE_WEEDS_NEEDED;
